+++ packages/psychart/src/psychart.ts
@@ -250,6 +250,32 @@ export class Psychart {
             });
     }
     /**
+     * Interpolate between "corner" psychrometric states.
+     */
+    private interpolate(states: PsyState[]): PsyState[] {
+        const data: PsyState[] = [states[0]];
+        for (let i = 1; i < states.length; i++) {
+            // Determine the start and end states.
+            const start: PsyState = states[i];
+            const end: PsyState = states[i - 1];
+            // Check if iso-rh, iso-wb, or iso-dp curved or straight lines
+            if (start.state.measurement === end.state.measurement && SMath.approx(start.state.other, end.state.other)) {
+                // Determine the dry bulb range
+                const minDb: number = SMath.clamp(Math.min(start.db, end.db), this.config.dbMin, this.config.dbMax);
+                const maxDb: number = SMath.clamp(Math.max(start.db, end.db), this.config.dbMin, this.config.dbMax);
+                // Compute several intermediate states with a step of `resolution`
+                for (let db: number = minDb; db <= maxDb; db += this.config.resolution) {
+                    data.push(new PsyState({ db: db, other: start.state.other, measurement: start.state.measurement }));
+                    // Stop generating if dew point exceeds maximum
+                    if (data[data.length - 1].dp > this.config.dpMax) {
+                        break;
+                    }
+                }
+            }
+        }
+        return data;
+    }
+    /**
      * Generate SVG path data from an array of psychrometric states.
      */
     private setPathData(path: SVGPathElement, psystates: PsyState[], closePath: boolean): void {
@@ -545,7 +571,16 @@ export class Psychart {
     /**
      * Draw a line between 2 arbitrary points on Psychart.
      */
-    public drawLine(start: Datum, end: Datum, colorHex: string, weight: number = 1): void {
+    public drawLine(start: Datum, end: Datum, colorHex: string, weight: number = 1, relHumType: DataOptions['relHumType'] = 'percent'): void {
+        // Hotfix: Adjust RH type
+        if (relHumType === 'percent') {
+            if (start.measurement === 'dbrh') {
+                start.other /= 100;
+            }
+            if (end.measurement === 'dbrh') {
+                end.other /= 100;
+            }
+        }
         const data: PsyState[] = [new PsyState(start)];
         // Check if iso-relative humidity (curved line)
         if (start.measurement === 'dbrh' && end.measurement === 'dbrh' && SMath.approx(start.other, end.other)) {